weird stuff
So I included a current tick fetch for both of the lift motors. This would be subtracted from the target goals from each motor in order to be able to reset the motor encoders as they are not being used, and reactivate them keeping there known position, for accurate movement.
Build successful, not sure how it would run, the display of the variable assignment seems strange.