A lot of stuff
A change in teleop to test for movement turning. An addition of full omni direction in auto move function. An addition of a power changing wait function that uses gyro to calculate how to change motor power to maintain heading. Finally a small addition of a  turn by gyro code at the and of any move function meant to correct after movement. Both straight wait and the final turn correction are commented out in order to test true omni movement prior to their implementation.